requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2.1
django-widget-tweaks==1.5.0
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.58.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
lxml==5.4.0
num2words==0.5.14
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zeep==4.3.1
zopfli==0.2.3.post1